+++ cayleypy/__init__.py
@@ -1,4 +1,5 @@
 from .cayley_graph import CayleyGraph
 from .graphs_lib import prepare_graph
 from .datasets import load_dataset
-from .bfs_special import bfs_bitmask, bfs_numpy
+from .bfs_bitmask import bfs_bitmask
+from .bfs_numpy import bfs_numpy
